nomads-agent/pyproject.toml
[project]
name = "nomads-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.68.2",
    "python-dotenv>=1.1.0",
]
